refactor: create package for crossplane related functions
